Explicit Int64 assignment for datetime
The datetime functions added to support numba in commit https://github.com/LimnoTech/HSPsquared/commit/e5d64a194eecd34bc7a6afeff0b5333fc2f52e05 required that integers input into these functions are 64bits or year information will be lost during bit manipulations. The previous implementation left integer type up to numba and in some instances could produce a in32 object. This commit causes integer conversion to be explicitly int64 so that year information is not lost.

+++ HSP2tools/readWDM.py
@@ -150,10 +150,10 @@ def readWDM(wdmfile, hdffile, compress_output=True):
 
 @njit 
 def _splitdate(x):
-    year = int(x >> 14)
-    month = int(x >> 10 & 0xF)
-    day = int(x >> 5 & 0x1F)
-    hour = int(x & 0x1F)
+    year = np.int64(x >> 14)
+    month = np.int64(x >> 10 & 0xF)
+    day = np.int64(x >> 5 & 0x1F)
+    hour = np.int64(x & 0x1F)
     return correct_date(year, month, day, hour, 0,0)
 
 @njit
